moved file location of file opener
